Ganache/requirements.txt
web3>=6.0.0
py-solc-x>=1.1.1